tests/wpt/tests/html/semantics/scripting-1/the-script-element/module/dynamic-import/propagate-nonce-inline-classic.html

<!DOCTYPE html>
<meta http-equiv="content-security-policy" content="script-src 'nonce-correct'">
<script nonce="correct" src="/resources/testharness.js"></script>
<script nonce="correct" src="/resources/testharnessreport.js"></script>
<script nonce="correct">
promise_test(t => {
  return import("./../imports-a.js").then(module => {
    assert_true(window.evaluated_imports_a);
    assert_equals(module.A["from"], "imports-a.js");
  });
}, "Dynamically imported module should eval when imported from script w/ a valid nonce.");
</script>
